Move new_transaction into eth.tools.factory.transaction
