+++ scripts/benchmark/checks/deploy_dos.py
@@ -24,6 +24,9 @@ from eth.constants import (
 from eth.rlp.blocks import (
     BaseBlock,
 )
+from eth.tools.factories.transaction import (
+    new_transaction
+)
 from _utils.chain_plumbing import (
     FUNDED_ADDRESS,
     FUNDED_ADDRESS_PRIVATE_KEY,
@@ -35,9 +38,7 @@ from _utils.compile import (
 from _utils.reporting import (
     DefaultStat,
 )
-from _utils.tx import (
-    new_transaction,
-)
+
 from .base_benchmark import (
     BaseBenchmark,
 )
