+++ tests/core/chain-object/test_build_block_incrementally.py
@@ -1,11 +1,10 @@
 import pytest
 
 from eth.chains.base import MiningChain
-from eth._utils.address import force_bytes_to_address
-
-from tests.core.helpers import (
-    new_transaction,
+from eth.tools.factories.transaction import (
+    new_transaction
 )
+from eth._utils.address import force_bytes_to_address
 
 
 ADDRESS_1010 = force_bytes_to_address(b'\x10\x10')
