+++ tests/core/helpers.py
@@ -1,7 +1,5 @@
 import sys
 
-from eth_utils.toolz import curry
-
 import pytest
 
 from eth_utils import (
@@ -10,50 +8,12 @@ from eth_utils import (
 )
 
 from eth.chains.base import MiningChain
-
-from eth.vm.spoof import (
-    SpoofTransaction,
-)
-
+from eth.tools.factories.transaction import new_transaction
 
 greater_equal_python36 = pytest.mark.skipif(
     sys.version_info < (3, 6),
     reason="requires python3.6 or higher"
 )
-
-
-@curry
-def new_transaction(
-        vm,
-        from_,
-        to,
-        amount=0,
-        private_key=None,
-        gas_price=10,
-        gas=100000,
-        data=b'',
-        chain_id=None):
-    """
-    Create and return a transaction sending amount from <from_> to <to>.
-
-    The transaction will be signed with the given private key.
-    """
-    nonce = vm.state.get_nonce(from_)
-    tx = vm.create_unsigned_transaction(
-        nonce=nonce,
-        gas_price=gas_price,
-        gas=gas,
-        to=to,
-        value=amount,
-        data=data,
-    )
-    if private_key:
-        if chain_id is None:
-            return tx.as_signed_transaction(private_key)
-        else:
-            return tx.as_signed_transaction(private_key, chain_id=chain_id)
-    else:
-        return SpoofTransaction(tx, from_=from_)
 
 
 def fill_block(chain, from_, key, gas, data):
