+++ tests/core/vm/test_vm_state.py
@@ -2,11 +2,11 @@ import pytest
 
 from eth_utils import ValidationError
 
+from eth.tools.factories.transaction import (
+    new_transaction
+)
 from eth.vm.interrupt import MissingAccountTrieNode
 
-from tests.core.helpers import (
-    new_transaction,
-)
 
 ADDRESS = b'\xaa' * 20
 OTHER_ADDRESS = b'\xbb' * 20
